Accelerate (a little bit) the default path for surface mesh drawing. We are still very slow, though:
- we should preprocess the mesh to draw large triangle/line strips
- we should store the normals (and NOT IN A SET!!)
- we should store the elements in a structure allowing a faster traversal
- we should do all this in single precision

The same remarks apply for the post-processing stuff. Worse, in
post-pro, due to Get_Coord and SaturateValues, we basically copy ALL
THE DATA SET at least one per drawing, even when we don't have to (no
Raise/Offset/Saturate)... This is horrible, and makes us lose all the
advantage of the fast list_PointerFast stuff I designed in the forst
place. Ahhh, the cost of generality...