This is a pretty big commit.
- The Numerical Recipes routines have been moved to their own
subdirectory: NR/

- These routines are now completely optional. Configure will
determine if they have to be compiled in (if it's a non-free version
of Gmsh) or if we should use the GSL routines instead.

- The GSL routines are there. Well, sort of. The SVD stuff works
great, but the minimization routines still need more work. The whole
infrastructure is there, though: we can already compile a fully GPL
compliant version of Gmsh by removing the NR subdirectory (or by
specifying --enable-gsl to configure).

- I also simplified the options passed to the compiler (OPTIM and
FLAGS).

That's all for today folks...