- Simplified the bounding box computation. It is now computed as follows (after each include/merge/open):
1. If there is a geometry (i.e., at least one geometrical point), the
bounding box is taken as the box enclosing all the geometrical points;

2. If there is no geometry but there is a mesh (i.e., at least one
mesh vertex), the bounding box is taken as the box enclosing all the
mesh vertices;

3. If there is no geometry and no mesh, but there are some
post-processing views, the bounding box is taken as the box enclosing
all the primitives of the last post-processing view.

This should fix the weird things that sometimes happen when we animate
scenes that do not contain any geometrical entities.

- Deprecated MergeWithBoundingBox()

- Removed Replot()

- Removed the "Last_NumberOfPoints" hack in the parser

- Removed CTX.expose