For all our mechanical engineering friends out there, I just cleaned up the stuff I did with Sean last year, i.e., allow to draw a vector field as a displacement map, with another scalar field displayed instead of the displacement norm.
It's basically the same as Plugin(DisplacementRaise), but it does not modify the
actual data inside the view (so e.g. you cannot save it). Instead, it's implemented
as an alternative "Vector display" mode. This allows to easily animate deformed meshed
with arbitrary scalar fields drawn on them. (The GUI for this is kinda lame, but
it works: just select Vector display->Raised scalar view, select the scalar view
number, et voila.)

I also generalized the smooth normal code path a little, so that we can have smooth
normals for vector-as-displacement and tensor-as-von-mises plots, too.