Removed the last bits of the "Discrete Surface/Line" stuff.
Gmsh can now simply read a .msh file, and use it as the boundary mesh
of a 3D mesh. For an example on how to do this, cf.
demos/sphere-discrete.geo.

This complements nicely the STL remeshing stuff: Gmsh can now
for example very easily be used to extrude surface meshes that
come from other tools.