Workaround for FLTK 1.1.7 popup menu bug (which could cause gmsh to crash when deleting a post-pro view).
Since we are going to make a LOT of changes in the geometry module in the
near future, I am going to create a branch with 1.64, which will be the
last "1.XX" version. I'll try to release 1.64 as soon as possible. Only
small bug fixes will go into that branch.

The trunk will get all the modeler changes, and will eventually become
Gmsh 2.0.