This is a big commit : do not hesitate to shout if a file is missing ;-)
The delquad algo is now much more stable and nice

Two 3D meshers have been added. 

One is an embryo of the 3D delhex mesh which is an extension of delquad.

The other one is MMG, the mobile mesh generator of CECILE DOBRZYNSKI (Bordeaux I) and PASCAL FREY (PARIS VI). We are very happy that this new anisotropic 3D mesher is now available as an open source !!