Allow to not compile the Gmsh executable (very usefull when using gmsh as a library)
Add a default implementation of fullMatrix<double>::invert and
fullMatrix<double>::determinant whn Lapack is not available. The algorirthms are bad but they allow to use polynomialBasis without installing BLAS and Lapack. The inverse is not computed a lot a time and the matrix is often not big in this case so there is no need of better algorithms