Implementation of gamepad for Windows. Now the use of gamepad is available for Linux and Windows. Still to be done for Apple.
Very usefull for shell model with compartments (boat, plane, building,...). Allows a good immersion into the model. I advise you to test.

Rendering seems to be better on Linux than on Windows: residual offsets on some axes signals are observed which leads to a possible little drifting of the model. Needs to shake joystick. It may come from the gamepad I used which is a old one.