Change default computation of mean plane (and thus of the parametrization) of plane surfaces, to make it more robust with respect to perturbations of the boundary.
This is crucial for sensitivity analyses and GFace::relocateMeshVertices(): after perturbation of the 
boundary, we want a parametrization of the surface that is "close" to the original one. This cannot be
guaranteed by the automatic SVD-based algorithm.

Unfortunately the new algorithm is also not guaranteed to work in all cases -- but unless we explicitely
prescribe the parametrization of plane surfaces (at least the normal) this is better.

I will revert this as soon as we have something better.