Addition of the Thin Layer plugin.
Requires the volume to be meshed with tetra, and remesh the surface, constraining vertices. After, the mesh can be cleaned and rebuilt, and the new constrained vertices will be used instead of the original ones. Constrained vertices appears in thin parts of the domain to ensure each vertex has a corresponding one on the other surface.