return "nothing" instead of last expression in Julia
It is conventional to return nothing when there
is no other sensible value to return. Before this commit
we implicitly returned the last expression, namely
ierr[1] != 0 which usually is false. This is a bit of a
"negative" return, especially since the comparison means the
operation was succesful.

This commits adds "return nothing" to the Julia API, when
the output argument list is empty.