API change: replaced setNodes/setElements by addNodes/addElements + added new mesh::clear()
Rationale: setNodes/setElements were intrinsically a bad design choice,
as these functions had to delete nodes and/or elements before storing the
new ones. Using these functions in the wrong way could lead to an invalid
mesh, with elements pointing to nodes that were deleted (in an adjacent
entity).

The only viable workflow is to delete the mesh (with the new clear()
function), before adding nodes/elements.