Add new gmsh::model::mesh::field::fromView function to API.
This function makes it easier to create a size field from an existing
view. This is used, for example, to set a view as a background mesh size
field. The existing setup in the API is somewhat tricky, since it requires users
to know about the correct field type ("PostView"), and the necessary
option ("IView").