enhanced non-recombined volume extrusion:  - use old (regular) tet splitting first, before using the tet split based on node nums  - make the tet split based on node tags deterministic (by comparing nums instead of ptrs)  - subdived all surfaces at once and make sure to handle dependencies to make surface mesh    consustent with volume mesh
One more enhancement is needed: "copied" surface meshes (the "top" surfaces) should also
be made conformal with the volume, in case the lateral side of an extrusion have themselves
been extruded.

This is the same issue as for transfinite volume meshes. We should thus simply implement
a surface "remesher" that extract the skin of the volume mesh.