Tentative fix for #1615.
The new implementation of buildConsecutiveListOfVertices() assumes
that the curve loops are correct. However, the loop calculation in
GEdgeLoop is purely topological, and can choose the "wrong" curve
on periodic surfaces in case several curves have the same
begin/end points, i.e. GEdgeLoop can chose a non-seam curve
starting on the "wrong" side of the parametrization.

We now use the original ordering of the OCC curves when possible,
and only recompute in case the loops are incorrect. The degenerate
"hack" in GEdgeLoop has also been removed.