Rewrote the Fortran API implementation
Use lists instead of strs to store the argument names and types
This is necesseary because a Fortran procedure/interface needs both
a dummy argument list and a variable declaration following the arg list.

This also make it easier to add argument `intent` and also provide an
interface for optional arguments.

Optional arguments could either be defined on the `interface` block
or create a module implementation which calls `present`.
I think in the former case the optional arguments would be handed to
the C backend so e.g. stderr will be reported through C not Fortran